@@ -1,6 +1,6 @@
 {
   "name": "esp-js-di-api",
-  "version": "5.1.2",
+  "version": "6.0.3-next.0",
   "private": true,
   "scripts": {
     "clean": "nps clean",
@@ -11,13 +11,13 @@
     "start": "./node_modules/.bin/ts-node src/index.ts"
   },
   "dependencies": {
-    "esp-js-di": "^5.1.2",
+    "esp-js-di": "^6.0.3-next.0",
     "prompt": "1.0.0"
   },
   "devDependencies": {
     "jest": "26.0.1",
     "ts-node": "^8.0.3",
-    "typescript": "3.9.5",
+    "typescript": "4.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "3.11.0"
