@@ -1,20 +1,20 @@
 {
   "name": "esp-js-react-agile-board",
-  "version": "5.1.2",
+  "version": "6.0.3-next.0",
   "description": "ESP JS React Components Sample.",
   "main": "src/app.js",
   "private": true,
   "dependencies": {
     "classnames": "2.2.6",
-    "esp-js": "^5.1.2",
-    "esp-js-react": "^5.1.2",
+    "esp-js": "^6.0.3-next.0",
+    "esp-js-react": "^6.0.3-next.0",
     "immutable": "^4.0.0-rc.12",
     "react": "16.13.1",
     "react-dom": "16.13.1"
   },
   "devDependencies": {
     "jest": "26.0.1",
-    "typescript": "3.9.5",
+    "typescript": "4.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "3.11.0"
