@@ -61,7 +61,7 @@
     "ts-node": "^8.6.2",
     "tslint": "^6.0.0",
     "tslint-loader": "^3.6.0",
-    "typescript": "3.9.5",
+    "typescript": "4.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "3.11.0",
