@@ -1,6 +1,6 @@
 {
   "name": "esp-js-polimer",
-  "version": "5.1.2",
+  "version": "6.0.3-next.0",
   "description": "Polimer adds redux-like reducers for esp-js",
   "keywords": [
     "state processor",
@@ -26,19 +26,19 @@
   "typings": ".dist/typings/index.d.ts",
   "license": "Apache-2.0",
   "dependencies": {
-    "esp-js": "^5.1.2",
-    "immer": "9.0.3",
+    "esp-js": "^6.0.3-next.0",
+    "immer": "9.0.7",
     "rxjs": "6.6.6"
   },
   "devDependencies": {
     "jest": "26.0.1",
-    "typescript": "3.9.5",
+    "typescript": "4.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "3.11.0"
   },
   "peerDependencies": {
-    "esp-js": "^5.0.0",
+    "esp-js": "^6.0.0",
     "immer": "^9.0.3",
     "rxjs": "^6.6.6"
   }
