@@ -1,6 +1,6 @@
 {
   "name": "esp-js-react",
-  "version": "5.1.2",
+  "version": "6.0.3-next.0",
   "description": "Evented State Processor (ESP) React Components",
   "keywords": [
     "esp",
@@ -24,21 +24,21 @@
     "dev": "nps dev"
   },
   "dependencies": {
-    "esp-js": "^5.1.2",
-    "esp-js-polimer": "^5.1.2",
+    "esp-js": "^6.0.3-next.0",
+    "esp-js-polimer": "^6.0.3-next.0",
     "prop-types": "^15.7.2",
     "react": "16.13.1"
   },
   "devDependencies": {
     "jest": "26.0.1",
-    "typescript": "3.9.5",
+    "typescript": "4.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "3.11.0"
   },
   "peerDependencies": {
-    "esp-js": "^5.0.0",
-    "esp-js-polimer": "^5.0.0",
+    "esp-js": "^6.0.0",
+    "esp-js-polimer": "^6.0.0",
     "prop-types": "^15.7.2",
     "react": "16.8.6"
   },
