@@ -1,6 +1,6 @@
 {
   "name": "esp-js-rx",
-  "version": "5.1.2",
+  "version": "6.0.3-next.0",
   "description": "ESP-JS-RX Package",
   "keywords": [
     "rxjs"
@@ -21,17 +21,17 @@
   "repository": "https://github.com/esp/esp-js",
   "devDependencies": {
     "jest": "26.0.1",
-    "typescript": "3.9.5",
+    "typescript": "4.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "3.11.0"
   },
   "dependencies": {
-    "esp-js": "^5.1.2",
+    "esp-js": "^6.0.3-next.0",
     "rxjs": "6.6.6"
   },
   "peerDependencies": {
-    "esp-js": "^5.0.0",
+    "esp-js": "^6.0.0",
     "rxjs": "^6.6.6"
   },
   "gitHead": "f6791a488f3fc5f59f20e19445b848d5e2f415c4"
