@@ -1,6 +1,6 @@
 {
   "name": "esp-js-ui-rxcompat",
-  "version": "5.1.2",
+  "version": "6.0.3-next.0",
   "description": "esp-js-ui-rxcompat Package",
   "main": ".dist/esp-js-ui-rxcompat.js",
   "types": "esp-js-ui-rxcompat.d.ts",
@@ -18,20 +18,20 @@
   },
   "devDependencies": {
     "jest": "26.0.1",
-    "typescript": "3.9.5",
+    "typescript": "4.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "3.11.0"
   },
   "dependencies": {
-    "esp-js": "^5.1.2",
-    "esp-js-rx": "^5.1.2",
+    "esp-js": "^6.0.3-next.0",
+    "esp-js-rx": "^6.0.3-next.0",
     "rxjs": "6.6.6",
     "rxjs-compat": "6.6.6"
   },
   "peerDependencies": {
-    "esp-js": "^5.0.0",
-    "esp-js-rx": "^5.0.0",
+    "esp-js": "^6.0.0",
+    "esp-js-rx": "^6.0.0",
     "rxjs": "^6.6.6",
     "rxjs-compat": "^6.6.6"
   }
