@@ -1,6 +1,9 @@
 export class ShellModuleContainerConst {
-    static get shell_model() { return 'shell_model'; }
+    static get workspace_model() { return 'workspace_model'; }
     static get workspace_region() { return 'workspace_region'; }
     static get blotter_region() { return 'blotter_region'; }
+    static get modal_region() { return 'modal_region'; }
     static get module_loader() { return 'module_loader'; }
+    static get app_preferences_model() { return 'app_preferences_model'; }
+    static get app_preferences_view_factory() { return 'app_preferences_view_factory'; }
 }
