@@ -21,6 +21,7 @@ import RegistrationModifier from './registrationModifier';
 import Guard from './guard';
 import EspDiConsts from './espDiConsts';
 import ResolverNames from './resolverNames';
+import {DefaultIsRegisteredQueryOptions} from './isRegisteredQueryOptions';
 
 export default class Container {
     constructor() {
@@ -33,6 +34,7 @@ export default class Container {
         this._resolvers = this._createDefaultResolvers();
         this._isDisposed = false;
         this._childContainers = [];
+        this._containerEventHandlers = new Map();
         this._registerSelf();
     }
     createChildContainer() {
@@ -89,18 +91,23 @@ export default class Container {
                 : InstanceLifecycleType.singleton
         };
         this._instanceCache[name] = instance;
+        this._raiseContainerEvent('instanceRegistered', name, instance)
     }
-    isRegistered(name) {
+    isRegistered(name, options = DefaultIsRegisteredQueryOptions) {
         this._throwIfDisposed();
         Guard.isNonEmptyString(name, 'Error calling isRegistered(name). The name argument must be a string and can not be \'\'');
-        var registration = this._registrations[name];
-        return !!registration;
+        const isRegistered = options.searchParentContainers
+            ? !!this._registrations[name]
+            : this._registrations.hasOwnProperty(name);
+        return isRegistered;
     }
-    isGroupRegistered(groupName) {
+    isGroupRegistered(groupName, options = DefaultIsRegisteredQueryOptions) {
         this._throwIfDisposed();
         Guard.isNonEmptyString(groupName, 'Error calling isGroupRegistered(groupName). The groupName argument must be a string and can not be \'\'');
-        var registration = this._registrationGroups[groupName];
-        return !!registration;
+        const isGroupRegistered = options.searchParentContainers
+            ? !!this._registrationGroups[groupName]
+            : this._registrationGroups.hasOwnProperty(groupName);
+        return isGroupRegistered;
     }
     resolve(name, ...additionalDependencies) {
         this._throwIfDisposed();
@@ -119,6 +126,7 @@ export default class Container {
             if (registration.instanceLifecycleType === InstanceLifecycleType.singleton || registration.instanceLifecycleType === InstanceLifecycleType.singletonPerContainer) {
                 this._instanceCache[name] = instance;
             }
+            this._raiseContainerEvent('instanceCreated', name, instance)
         } else if(additionalDependencies.length > 0) {
             throw new Error("The provided additional dependencies can't be used to construct the instance as an existing instance was found in the container");
         }
@@ -145,6 +153,30 @@ export default class Container {
         Guard.isNonEmptyString(name, 'Error calling addResolver(name, resolver). The name argument must be a string and can not be \'\'');
         Guard.isNotNullOrUndefined(resolver, `Error calling addResolver(name, resolver). Provided resolver for [${name}] can not be null or undefined`);
         this._resolvers[name] = resolver;
+    }
+    on(eventType, eventHandler) {
+        Guard.isNonEmptyString(eventType, 'Error calling on(eventType, eventHandler). The eventType argument must be a string and can not be \'\'');
+        Guard.isFunction(eventHandler, 'Error calling on(eventType, eventHandler). The eventHandler argument must be a function can not be null or undefined');
+        let handlers = this._containerEventHandlers.get(eventType);
+        if (!handlers) {
+            handlers = [];
+            this._containerEventHandlers.set(eventType, handlers);
+        }
+        let handlerExists = handlers.some(handler => handler === eventHandler);
+        Guard.isFalsey(handlerExists, 'Error calling on(eventType, eventHandler). The eventHandler passed is already registered');
+        handlers.push(eventHandler);
+    }
+    off(eventType, eventHandler){
+        Guard.isNonEmptyString(eventType, 'Error calling off(eventType, eventHandler). The eventType argument must be a string and can not be \'\'');
+        Guard.isFunction(eventHandler, 'Error calling off(eventType, eventHandler). The eventHandler argument must be a function can not be null or undefined');
+        let handlers = this._containerEventHandlers.get(eventType);
+        if (!handlers) {
+            return;
+        }
+        const removeAtIndex = handlers.indexOf(eventHandler);
+        if (removeAtIndex > -1) {
+            handlers.splice(removeAtIndex, 1);
+        }
     }
     dispose() {
         this._disposeContainer();
@@ -281,6 +313,22 @@ export default class Container {
     _registerSelf() {
         // register the child with itself so any dependency that wants to resolve a container get's the container at the same scope as itself (i.e. the container that built it).
         this.registerInstance(EspDiConsts.owningContainer, this);
+    }
+    _raiseContainerEvent(eventType, name, instance) {
+        let handlers = this._containerEventHandlers.get(eventType);
+        if (!handlers) {
+            return;
+        }
+        let notification = {
+            name,
+            instance,
+            eventType
+        };
+        // copy the list as we're about to call external code which could be reentrant
+        let copy = handlers.slice(0);
+        for (let i = 0; i < copy.length; i ++) {
+            copy[i](notification)
+        }
     }
     _throwIfDisposed() {
         if (this._isDisposed) throw new Error("Container has been disposed");
