@@ -1,17 +1,15 @@
 // import for side effects
 import './modelBuilder';
 
-export {multipleEvents, PolimerEventHandler, PolimerHandlerMap} from './stateEventHandlers';
+export {PolimerEventHandler} from './stateEventHandlers';
 export {PolimerModel} from './polimerModel';
 export {ImmutableModel} from './immutableModel';
 export {PolimerModelBuilder} from './modelBuilder';
 export {
     InputEvent,
     InputEventStream,
-    OutputEventStreamFactory,
     OutputEvent,
     OutputEventStream,
-    InputEventStreamFactory,
     eventTransformFor
 } from './eventTransformations';
 export {PolimerEvents} from './polimerEvents';
