@@ -2,8 +2,7 @@ import {Container, Resolver} from 'esp-js-di';
 import {Guard} from 'esp-js';
 
 /**
- * A
- * resolver that allows you to to inject/resolve a literal value (for example a string)
+ * A resolver that allows you to inject/resolve a literal value (for example a string)
  * @deprecated this is now included by default in esp-js-di. 
  * This can still be added to a container, it will just replace the build in one with the same implementation. 
  */
